float stocktaking

+++ resources/views/stocktakings/monthly/audit_1.blade.php
@@ -426,7 +426,7 @@
 
 						$('#confirm').show();
 						$('#progress-bar').addClass('active');
-										
+
 					}else{
 						body += '<td '+css+'>-</td>';
 					}
@@ -641,7 +641,7 @@
 		lot_uom = (lot_uom || 0);
 		z1 = (z1 || 0);
 
-		var total = (parseInt(lot) * parseInt(lot_uom)) + parseInt(z1);
+		var total = (parseFloat(lot) * parseFloat(lot_uom)) + parseFloat(z1);
 		document.getElementById("total").value = total;
 	}
